Fix --no-git argument resolution.
The --no-git argument was superfluous as it was set to True
by default.
This commit flips this, and assumes
git is not available, and will check and update only if the user
asks for it (--no-git == False).
To make this more clear I have renamed the variable to
should_init_git.